apply SVN commit 5432 to Git

+++ src/Gui/View3DPy.cpp
@@ -419,7 +419,7 @@ Py::Object View3DInventorPy::viewRotateLeft(const Py::Tuple& args)
       SbRotation rot = cam->orientation.getValue();
       SbVec3f vdir(0, 0, -1);
       rot.multVec(vdir, vdir);
-      SbRotation nrot(vdir, M_PI/2);
+      SbRotation nrot(vdir,float( M_PI/2));
       cam->orientation.setValue(rot*nrot);
     }
     catch (const Base::Exception& e) {
@@ -445,7 +445,7 @@ Py::Object View3DInventorPy::viewRotateRight(const Py::Tuple& args)
       SbRotation rot = cam->orientation.getValue();
       SbVec3f vdir(0, 0, -1);
       rot.multVec(vdir, vdir);
-      SbRotation nrot(vdir, -M_PI/2);
+      SbRotation nrot(vdir, float(-M_PI/2));
       cam->orientation.setValue(rot*nrot);
     }
     catch (const Base::Exception& e) {
